Refactor producer task queue.

+++ elastic-job-cloud/elastic-job-cloud-master/src/main/java/com/dangdang/ddframe/job/cloud/producer/TaskProducerSchedulerRegistry.java
@@ -22,26 +22,22 @@ import com.dangdang.ddframe.job.cloud.config.ConfigurationService;
 import com.dangdang.ddframe.reg.base.CoordinatorRegistryCenter;
 import com.google.common.base.Optional;
 
-import java.util.concurrent.ConcurrentHashMap;
-
 /**
  * 发布任务作业调度注册表.
  *
- * @author zhangliang
+ * @author caohao
  */
 public class TaskProducerSchedulerRegistry {
     
     private static volatile TaskProducerSchedulerRegistry instance;
     
-    private final ConcurrentHashMap<String, TaskProducerScheduler> taskProducerSchedulerMap = new ConcurrentHashMap<>(65535);
+    private TaskProducerScheduler schedulerInstance;
     
-    private final CoordinatorRegistryCenter registryCenter;
-    
-    private final ConfigurationService configService;
+    private ConfigurationService configService;
     
     private TaskProducerSchedulerRegistry(final CoordinatorRegistryCenter registryCenter) {
-        this.registryCenter = registryCenter;
         configService = new ConfigurationService(registryCenter);
+        schedulerInstance = new TaskProducerScheduler(registryCenter);
     }
     
     /**
@@ -62,12 +58,10 @@ public class TaskProducerSchedulerRegistry {
     }
     
     /**
-     * 将注册中心中的云作业配置注册调度.
+     * 启动作业调度器.
      */
-    public void registerFromRegistryCenter() {
-        for (CloudJobConfiguration each : configService.loadAll()) {
-            register(each);
-        }
+    public void startup() {
+        schedulerInstance.startup(configService.loadAll());
     }
     
     /**
@@ -76,20 +70,13 @@ public class TaskProducerSchedulerRegistry {
      * @param jobConfig 作业配置
      */
     public void register(final CloudJobConfiguration jobConfig) {
-        String jobName = jobConfig.getJobName();
-        if (taskProducerSchedulerMap.containsKey(jobName)) {
-            taskProducerSchedulerMap.get(jobName).shutdown();
-            taskProducerSchedulerMap.remove(jobName);
-        }
-        Optional<CloudJobConfiguration> jobConfigFromZk = configService.load(jobName);
+        schedulerInstance.register(jobConfig);
+        Optional<CloudJobConfiguration> jobConfigFromZk = configService.load(jobConfig.getJobName());
         if (!jobConfigFromZk.isPresent()) {
             configService.add(jobConfig);
         } else if (!jobConfigFromZk.get().equals(jobConfig)) {
             configService.update(jobConfig);
         }
-        TaskProducerScheduler taskProducerScheduler = new TaskProducerScheduler(jobConfig, registryCenter);
-        taskProducerScheduler.startup();
-        taskProducerSchedulerMap.put(jobName, taskProducerScheduler);
     }
     
     /**
@@ -98,10 +85,17 @@ public class TaskProducerSchedulerRegistry {
      * @param jobName 作业名称
      */
     public void deregister(final String jobName) {
-        if (taskProducerSchedulerMap.containsKey(jobName)) {
-            taskProducerSchedulerMap.get(jobName).shutdown();
-            taskProducerSchedulerMap.remove(jobName);
+        Optional<CloudJobConfiguration> jobConfigFromZk = configService.load(jobName);
+        if (jobConfigFromZk != null && jobConfigFromZk.isPresent()) {
+            schedulerInstance.deregister(jobConfigFromZk.get());
+            configService.remove(jobName);
         }
-        configService.remove(jobName);
+    }
+    
+    /**
+     * 关闭作业调度器.
+     */
+    public void shutdown() {
+        schedulerInstance.shutdown();
     }
 }
